chore: Updated `<MediaCard />` design
- Also updated `formatForMediaCard()`
