+++ mobile/src/modules/media/components/index.ts
@@ -1,3 +1,4 @@
+export { MediaCard, MediaCardPlaceholderContent } from "./MediaCard";
 export {
   NextButton,
   PlayToggleButton,
